회원의 프로필 사진 추가(Ajax 작성 중,,)

+++ src/test/java/org/choongang/email/EmailApiTest.java
@@ -14,22 +14,24 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class EmailApiTest {    // API 통합 테스트
+public class EmailApiTest {
     @Autowired
-    private MockMvc mockMvc ;
+    private MockMvc mockMvc;
 
     @Test
     @DisplayName("이메일 인증 코드 발급 및 검증 테스트")
     void sendVerifyEmailTest() throws Exception {
-        // 인증 코드 발급 테스트
+        /* 인증 코드 발급 테스트 S */
         HttpSession session = mockMvc.perform(get("/api/email/verify?email=2sujg97@naver.com"))
                 .andDo(print())
                 .andExpect(status().isOk())
-                .andReturn().getRequest().getSession() ;
-        Integer authNum = (Integer) session.getAttribute("EmailAuthNum") ;
+                .andReturn().getRequest().getSession();
+        Integer authNum = (Integer)session.getAttribute("EmailAuthNum");
+        /* 인증 코드 발급 테스트 E */
 
-        // 인증 코드 검증 테스트
-        mockMvc.perform(get("/api/email/auth_check?authNum=" + authNum))
-                .andDo(print()) ;
+        /* 인증 코드 검증 테스트 S */
+        mockMvc.perform(get("/api/email/auth_check?authNum=" + authNum.intValue()))
+                .andDo(print());
+        /* 인증 코드 검증 테스트 E */
     }
 }
